requirements.txt
Django==2.2.2
python-dotenv
psycopg2==2.8.4
dj-database-url==0.4.2 
requests
django-cors-headers==3.0.2
djangorestframework==3.10.3
pyjwt
requests
